cluster: add get_guest_config_properties
Akin to the singular get_guest_config_property, but with the
possibility to query a list of properties.

Uses the CFS_IPC_GET_GUEST_CONFIG_PROPERTIES introduced in the
previous patch.

Note that the same details apply w.r.t. parsing and permissions as
the singular variant, iow. one needs to take caution and filter
allowed guests views  on call site when using this.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
 [T: reword/extend commit message ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>