datacenter config: mark notification settings as deprecated/unused
These were part of the first version of the notification overhaul
which was already rolled out in pvetest. To avoid breakage for users
who may have used the version from pvetest, we do not remove them yet
and only mark them as unused or deprecated. They can be removed at
some point in the future.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>