cfs: add 'sdn/pve-ipam-state.json' to observed files
We want to migrate from the unnecessary restricted ipam.db file inside
the root-only priv/ directory as that is reserved for secrets that
allow access to (external) resources with high exposure.
The ipam.db was for internal state, while that is somewhat sensitive
information it's not more so than the raw configuration files for
storage, guests, ha and so on, and all those files are not directly
accessible to the user as we always have the API in front of the
pmxcfs anyway.

The background here is that moving that file out of priv/ allows
non-root daemons like pveproxy to read it, which is required for some
firewall related development. And in general we want to go more in the
direction of less stuff running as root, not more stuff.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>