pmxcfs: add clang-format config add make targets to auto-format code
We are in the process of auto-formatting our Perl (using perltidy) and
JavaScript (using biome) code bases, just like we adopted rustfmt
already a few years ago. So lets also format our C code bases while at
it, while they see less activity compared to perl/js/rust at the
moment, there is still some development going on and using
clang-format works well with a rather minimal config, so just go for
it.

Add a make 'format' target with an 'tidy' alias as we use that for the
perl stuff (mostly due to the tool being named perltidy there). Using
format as main target is chosen in a similar spirit, as for clang the
tool is named clang-format and there is a clang-tidy which is a linter
though.

Check in a .clang-format config here, basing of the LLVM style but
with 4 spaces as indentation and up to 100 character columns text
width, and also try to avoid packing arguments to much into less
lines, this hurts readability and line-based editing can be nicer
w.r.t conflicts and blame/history tracking in git anyway.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>