rrd: cope with only partially renamed memavailable field
During initial development of the revised metrics series this field
was sourced by 'memfree' from /proc/meminfo as parsed by our ProcFS
tools, but as memavailable got added for PVE 8.4, as it much better
shows the state of actually available memory [0], I suggested to Aaron
to move the new RRD format also to that metric. While that was done,
the RRD definition was still using the old name, while I fixed in in
pmxcfs, I forgot to do so in the upgrade migration tool, which then
created an RRD file with the old name.

Thus map the old name to the new one here in case there is no existing
entry for the new one already, if one would be parsed later it would
override this fallback anyway.

[0] as while the kernel tries to use unused memory to improve
performance, it has no issue to release that again quickly on memory
pressure.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>