rrd metrics: fix resolution and point count for month and year time frames
The month time frame for the new RRD files uses 1440 data points with
a resolution of 30 minutes, not 121 points of 6 hour length.

The year-period also had a typo, the amount of data points 1140 should
have been 1440 points.

This improves the returned and displayed metrics in the API/UI
especially noticeable for the month time-frame due to having a 12x
higher resolution now.

Fixes: 1a483ba ("rrd: adapt to new RRD format with different aggregation windows")
Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>
Reviewed-by: Laurențiu Leahu-Vlăducu <l.leahu-vladucu@proxmox.com>
Tested-by: Laurențiu Leahu-Vlăducu <l.leahu-vladucu@proxmox.com>
Link: https://lore.proxmox.com/20251103220024.2488005-2-a.lauterer@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>