pmxcfs: status: always use 9.0 rrd files
Assuming that we can keep the old RRD files around and fetch data
directly there if needed (see previous commit), we can skip the
migration step and just write to new 9.0 RRD files.

This simplifies the whole approach introduced in commit 9b00ac9
status: introduce new pve-{type}- rrd and metric format

We will always create a new 9.0 file if not present and then need to
handle the following cases:

* old pve2-{type} format -> pad missing columns
* pve-{type}-9.0 matches -> no changes needed
* pve-{type}- received, but version doesn't match -> new format, cut at
  expected columns

Since we don't need the old RRD definitions anymore, they are removed.
Otherwise the compile will throw an unused variable warning.
The `rrd_def_{type}_columns` variables get a `_pve2` suffix to make it
clear what they are for.

The `checked_mkdir` pattern for VMs is aligned with how we do it for
nodes and storages, to call it once more, to be safe.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>
Reviewed-by: Laurențiu Leahu-Vlăducu <l.leahu-vladucu@proxmox.com>
Tested-by: Laurențiu Leahu-Vlăducu <l.leahu-vladucu@proxmox.com>
Link: https://lore.proxmox.com/20251103220024.2488005-4-a.lauterer@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>